app/requirements.txt
litserve
langchain_community
sentence-transformers
pydantic
python-dotenv
langchain_huggingface
transformers
torch